Add a 256 colors version of the colorscheme
@scottchiefbaker proposed a 256 colors version of Gotham in #1, but I didn't
feel like losing the accuracy of the tailored 16 colors against the portability
and ease of use (no terminal emulator colorscheme needed) of the 256 colors.

This commit adds a gotham-256 version of the colorscheme so that everyone is
able to choose what version to use.

There are a ton of things that could be extracted (maybe into `autoload/`), but
for now I'm leaving this untouched in order to avoid extra boot time when
loading an external file in `autoload/`.