Highlight the status line properly
Close #9

I thought the foreground color of StatusLine and StatusLineNC impacted the space
in the status line between two vertical splits. I was wrong; that's the
background color.

Before this commit, the status line in an inactive window was "invisible" (same
bg and fg).  I made the default status line a little brighter and changed the
fg of the inactive status line so that now it works like it should.